# Update queried metric to determine tps in system tests